move organization loader tests
