+++ api-js/src/loaders/organizations/__tests__/load-organization-connections-by-domain-id.test.js
@@ -5,14 +5,14 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
 const {
   orgLoaderConnectionArgsByDomainId,
   orgLoaderByKey,
-} = require('../loaders')
+} = require('../..')
 
 describe('given the load organizations connection function', () => {
   let query,
